LayoutTests/media/modern-media-controls/tracks-support/ios/tracks-support-off-text-track.html

<!DOCTYPE html>
<script src="../../resources/media-controls-utils.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
<script src="../../../../resources/js-test-pre.js"></script>
<body>
<video src="../../../content/CC+Subtitles.mov" style="position: absolute; left: 0; top: 0; width: 300px;" controls autoplay muted playsinline></video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Ensure we have the Off text track selected by default.");

const media = document.querySelector("video");
const shadowRoot = window.internals.shadowRoot(media);

media.addEventListener("play", () => {
    media.pause();

    shouldBecomeDifferent("shadowRoot.querySelector('button.tracks')", "null", () => {
        shouldBecomeDifferent("shadowRoot.querySelector('button.tracks').getBoundingClientRect().width", "0", async () => {
            debug("Tapping tracks button...");
            let [contextmenu] = await Promise.all([
                getTracksContextMenu(),
                pressOnElement(shadowRoot.querySelector("button.tracks")),
            ]);
            debug(JSON.stringify(contextmenu, null, 2));

            media.remove();
            finishJSTest();
        });
    });
});

</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
