LayoutTests/media/modern-media-controls/tracks-support/ios/tracks-support-off-text-track-expected.txt

Ensure we have the Off text track selected by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
Tapping tracks button...
[
  {
    "children": [
      {
        "title": "Languages",
        "children": [
          {
            "title": "English Sound",
            "checked": true
          },
          {
            "title": "French Sound"
          },
          {
            "title": "Spanish Sound"
          }
        ]
      },
      {
        "title": "Subtitles",
        "children": [
          {
            "title": "Off",
            "checked": true
          },
          {
            "title": "Auto (Recommended)"
          },
          {
            "title": "English Subtitles"
          },
          {
            "title": "French Subtitles"
          },
          {
            "title": "German Subtitles"
          },
          {
            "title": "Spanish Subtitles"
          }
        ]
      }
    ]
  }
]
PASS successfullyParsed is true

TEST COMPLETE
